Audio output.
This is 2 channel audio output. Latency appears to be ~50ms, but we might be able to push that lower.

Add a few data structures for lockfree/kernel-call-free coding. We use more semaphores than I'd like, but way less than my previous attempts at this.

Also turn on all the warnings with -Wall and -Werror, which found bugs in other things.