Introduce PannerBank and convert HrtfConvolver to HrtfPanner.
For efficiency, Synthizer processes panners as groups of 4 with SIMD, then a
PannerBank performs the final summation at the end, paying the SIMD-unfriendly
horizontal adds only once.  For future-proofing, we don't assume that it is
always 4 lanes, but instead make the PannerBank smart enough to handle
heterogeneous panners in future for i.e. surround sound handling.