Introduce a context object and related machinery:
This adds a context object, which will be the entrypoint to the library once
the C API is exposed.

invokable.hpp provides a very
lightweight mechanism for enqueueing cross-thread tasks, like a combination of
std::promise and std::future with the additional limitation that the invokables
need to be kept alive by whoever is managing them.  Most of the interaction
with context goes through an invokable, on the audio thread.  The
Synthizer approach to avoiding locking problems is to wake this thread on
either an invokable or the need for a block of audio, then check whether audio
is needed first.

Finally, add a pregenerate callback mechanism, which lets us move wav_test.cpp
to the context interface without having yet created sources as an abstraction.
Pregenerate callbacks will be removed from the library as soon as the higher
level pieces exist and are not a completed mechanism.