add a C API.
Note to early adopters: this is probably broken due to lifetime issues;
if you encounter a reliable way to crash it around syz_handleDecRef and/or at
program exit, please open an issue with example code.

Also need to nail down some thread safety and logging is very slow if enabled.