Refactor the property enums to allow simulating dynamic languages in C
This is now one enum, with constants of the form `SYZ_P_XXX`, which is both
shorter and opens us up to the following minor advantage: C code can now write
somewhat polymorphic code, i.e. manipulating the context and source position
properties with the same functions, or seeking any kind of generator with
a position property, without having to branch on the type.

New constants will be added to the bottom of the enum as needed. The canonical
source for what objects have what properties is synthizer_properties.h, but
will transition to the manual shortly.