Implement a router, except for the part where we actually do the routing.
This scafolds the ability to have single-level routing backed by a sorted
vector. Routes suppor fadein and fadeout functionality, as well as smoothe gain
changes at runtime.

The actual routing function and integration with the rest of synthizer is still
to be done; as of this commit this is still nonfunctional (but building) code.