Move to a simpler effects algorithm that doesn't use finite state machines
This is done by realizing that there is no difference between a route that's
not crossfading which also happens to have a gain of 0.0, and a dead route. By
pulling out the fading logic to a LinearFader class that we can also reuse
later with a couple of particular guarantees, the entire thing is just one
fader.