All audio generation happens in the miniaudio thread
This has the cost of making property writes even slower, but is necessary to
work on Linux (as well as potentially other platforms).  This is because audio
thread priority and scheduling turns out to be more complicated than doing
nothing on anything but Windows.

We document that property writes are slower in the release notes, but
users have already been clearly told to only use them for debugging purposes,
so I don't consider this a breaking change. I'm hoping to be able to do more
property reading work before 1.0, but I haven't scoped the effort yet.

This is also the first step toward not having a global output device, which is
in theory a good idea for Windows but nonetheless not helpful from a code
perspective.

Refs #1 because Linux stability work