Expose `syz_handleGetObjectType` for querying the type of an object at runtime
As the event system continues to come online, it will become more and more
necessary to query the type of a handle being as some events may be generated
by more than one object type (e.g. BufferGenerator and StreamingGenerator both
fire finished).  Whether this will see heavy use, who knows, but it's a useful
and trivial interface to add.

Refs #21