Better HRTF implementation
Fix the minimum phase conversion, improve the behind/in front difference with
a lowpass, remove the base frequencies. Generally clean up and debug mit.py
into something that's maintainable.  It was last thrown together before the
library even existed as a quick proof of concept.  Also, increase the fft length
for the magnitude responses to match the sample rate; this gives much better
processing when averaging power and so on.

Refs #27. Doesn't close it because we want to collect feedback.