Introduce a markPropertyUnchanged interface
We won't be changing the property system for a long time if ever, and if we
do the entire thing will need to be rewritten from the ground up anyway. Bow to
the inevitable, and adopt the old acquireProperty hack behind a proper
interface for cases where properties need to be seen as unchanged in the first
audio tick.

Fixes #53