Increment/decrement event handle reference counts to guarantee validity of handles for the duration of event processing.
And add a flag so that users may conveniently request that they take ownership
of the handles instead of them being freed at `syz_eventDeinit` time.

This solves the lifetime issue in #21 and makes events safe to use in
multithreaded contexts.  For bindings such as Rust, it is now possible to
put handles behind smart pointer abstractions and support cloning by taking on
the burden of translating event handles out, without having to e.g. figure out
typesafe ways to keep the user from "stealing" event handles (for bindings like
Python where preserving object tidentity is important, only the "this handle
definitely lives long enough" concern applies).

Doc updates will be handled as part of #28 in one go in what is essentially
going to be a rewrite of most of the manual.