Support decoding from libsndfile
This requires the user to specify a path to the dll/shared object, and can only
work on streams that support seeking due to inrternal libsndfile
limitations.  We now support all of the formats libsndfile supports, provided
that the user distributes and enables it.

Fixes #37 because we now support ogg, and also mostly fixes other format
requests we might get in future.  That said, libsndfile has a lot of corner
cases, doesn't work with streaming, and is at least somewhat slow, so users are
encouraged to use the built-in formats and leave it disabled when possible.