Set property values when they have timelines
Since this is just block level, we pretend that the user set the property and
drop their value if they added one.  We will introduce support for relative
timelines later, which is more complicated: we must evaluate late in that
case, which e.g. breaks the acquisition bit clearing logic.  Avoid the bugs,
and implement the simplest version that gets us functionality.

Refs #73