All automation timing is context-relative.
And advances when anything but the context is paused.  This is because a
trade-off must be made: we can either allow for synchronizing automation of
objects together so that they can be in sync, or we can allow for automation
on a generator to be relative to zero and sample-accurate at the beginning.
The former is better for games or anything where automation accuracy on the
order of 10s of ms is fine, and the latter better for music.  Since this is
primarily for games/VR, we go ahead and make the trade-off that allows global
object synchronization.

The difficulty here is that adding a switch between the modes is nontrivial,
since faders aren't able to just suddenly jump ahead because the user changed
their mind.  Or to go backward for that matter.  If it turns out to be the case
that demand for keyboard-level quality is desired, we can revisit this then:
it's not impossible to do, and things like WebAudio-style "start paused, then
play when I tell you to" are doable.

Refs #73