FastSineBank: an abstraction for running sine banks which doesn't require actually running trigonometry.
On X86 you can't use fsin due to accuracy and even if you can it's very
expensive.  But there's a trick: use the sum of angle formulas from trig, and
the entire thing can become some multiplications and additions.  This is not
accurate over the long run, so synchronize with the real values once per block.

Will be used to make a `SineBankGenerator`, which sums sine waves for
chiptune-quality waveforms.

Refs #17